+++ avian2d/src/lib.rs
@@ -17,7 +17,6 @@ use bevy_tnua_physics_integration_layer::data_for_backends::{
     TnuaProximitySensor, TnuaProximitySensorOutput, TnuaRigidBodyTracker, TnuaSensorOf, TnuaToggle,
 };
 use bevy_tnua_physics_integration_layer::math::*;
-use bevy_tnua_physics_integration_layer::subservient_sensors::TnuaSubservientSensor;
 pub use spatial_ext::TnuaSpatialExtAvian2d;
 
 use bevy_tnua_physics_integration_layer::*;
@@ -79,28 +78,7 @@ impl Plugin for TnuaAvian2dPlugin {
             self.schedule,
             apply_motors_system.in_set(TnuaPipelineSystems::Motors),
         );
-        app.add_systems(
-            Update,
-            ensure_subservient_sensors_are_colliders_of_their_bevy_parents,
-        );
-        app.register_required_components::<TnuaSubservientSensor, Position>();
-        app.register_required_components::<TnuaSubservientSensor, Rotation>();
         app.register_required_components_with::<TnuaGravity, GravityScale>(|| GravityScale(0.0));
-    }
-}
-
-#[allow(clippy::type_complexity)]
-fn ensure_subservient_sensors_are_colliders_of_their_bevy_parents(
-    query: Query<(Entity, &ChildOf), (With<TnuaSubservientSensor>, Without<ColliderOf>)>,
-    mut commands: Commands,
-) {
-    for (entity, child_of) in query.iter() {
-        commands
-            .entity(entity)
-            // NOTE: Use the parent from child_of instead of `TnuaSubservientSensor::owner_entity`
-            // because Bevy parenting is how it works in Rapier and the sensor's owner is allowed
-            // to be different than the Bevy parent.
-            .insert_if_new(ColliderOf { body: child_of.0 });
     }
 }
 
@@ -160,7 +138,6 @@ fn update_proximity_sensors_system(
         &Rotation,
         Option<&Collider>,
         Option<&TnuaAvian2dSensorShape>,
-        Option<&TnuaSubservientSensor>,
         Option<&TnuaToggle>,
     )>,
     collision_layers_query: Query<&CollisionLayers>,
@@ -180,7 +157,7 @@ fn update_proximity_sensors_system(
 ) {
     sensor_query.par_iter_mut().for_each(
         |(mut sensor, &TnuaSensorOf(owner_entity), shape, mut ghost_sensor)| {
-            let Ok((position, rotation, collider, owner_shape, subservient, tnua_toggle)) =
+            let Ok((position, rotation, collider, owner_shape, tnua_toggle)) =
                 owner_query.get(owner_entity)
             else {
                 return;
@@ -211,12 +188,6 @@ fn update_proximity_sensors_system(
                 // than it should be.
                 normal: Dir3,
             }
-
-            let owner_entity = if let Some(subservient) = subservient {
-                subservient.owner_entity
-            } else {
-                owner_entity
-            };
 
             let collision_layers = collision_layers_query.get(owner_entity).ok();